security(Scopes): Addess security vulnerabilities related to scopes
1. When using Client Credentials any user with the shared credentials for this client would be able to gain the administrator role. (This was not originally reported here.)
2. After issuing a valid token, if the roles assigned to the user or the client changed the access tokens previously issues could grant the original roles.

In order to exploit vulnerability #1 a user would need to have access to the client's id and password and they should be using the client credentials grant. Nonetheless this constitutes a privilege escalation issue, I dim it to be `Critical`.

Vulnerability #2 is mitigated by the fact that there needs to be a well timed chain of events. The exploit would only be accessible during the time after an special edit (remove a role) was made and the token was expired. In addition to that a token using that particular role being removed would have to exist prior to the edit. Nonetheless this constitutes a privilege escalation issue, I dim it to be `Less Critical`.